new commit

+++ print_address.c
@@ -14,7 +14,7 @@ int print_address(va_list args)
     int chars_printed = 0, temp_index = 0, digit, i;
     char hex_digits[16] = "0123456789abcdef";
     char temp_buffer[32];
-	unsigned long long int value;
+	unsigned long int value;
 
     if (ptr == NULL)
         return (_printf("(null)"));
@@ -23,7 +23,7 @@ int print_address(va_list args)
     chars_printed += _putchar('x');
 
 
-    for (value = (unsigned long long int)ptr; value > 0; value /= 16) 
+    for (value = (unsigned long int)ptr; value > 0; value /= 16) 
 	{
         digit = value % 16;
         temp_buffer[temp_index++] = hex_digits[digit];
